# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Backend Nutrisionist API",
    "version": "1.0.0",
    "description": "API for nutrisionist."
  },
  "servers": [
    {
      "url": "https://backend-nutrisionist-production.up.railway.app/api/v1",
      "description": "Deploy"
    },
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Local server"
    }
  ],
...
  "paths": {
...
    "/auth": {
      "get": {
        "summary": "Get current logged-in user data",
        "tags": ["Users"],
        "operationId": "getCurrentUser",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "User data retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Success"
                    },
                    "message": {
                      "type": "string",
                      "example": "User data retrieved successfully"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "username": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string"
                        },
                        "dob": {
                          "type": "string"
                        },
                        "gender": {
                          "type": "string"
                        },
                        "avatar": {
                          "type": "string",
                          "format": "uri"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "No token provided"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/auth/update": {
      "put": {
        "summary": "Update current user profile",
        "tags": ["Users"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "operationId": "updateUser",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "dob": {
                    "type": "string",
                    "format": "date"
                  },
                  "gender": {
                    "type": "string",
                    "enum": ["male", "female", "other"]
                  },
                  "avatar": {
                    "type": "string",
                    "format": "uri"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Success"
                    },
                    "message": {
                      "type": "string",
                      "example": "User updated successfully"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "username": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string"
                        },
                        "dob": {
                          "type": "string"
                        },
                        "gender": {
                          "type": "string"
                        },
                        "avatar": {
                          "type": "string",
                          "format": "uri"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "No token provided"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/physical-data": {
      "get": {
        "summary": "Retrieve physical data for the current user",
        "tags": ["Physical Data"],
        "operationId": "getPhysicalDatas",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Physical data retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "example": "Success" },
                    "message": {
                      "type": "string",
                      "example": "Physical data was successfully retrieved"
                    },
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/PhysicalData" }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "summary": "Create new physical data",
        "tags": ["Physical Data"],
        "operationId": "createPhysicalData",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PhysicalDataInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Physical data created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "example": "Success" },
                    "message": {
                      "type": "string",
                      "example": "Data successfully created"
                    },
                    "data": { "$ref": "#/components/schemas/PhysicalData" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Incomplete data"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/physical-data/{id}": {
      "put": {
        "summary": "Update physical data by ID",
        "tags": ["Physical Data"],
        "operationId": "updatePhysicalData",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Physical data ID",
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PhysicalDataInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Physical data updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "example": "Success" },
                    "message": { "type": "string", "example": "Data updated" },
                    "data": { "$ref": "#/components/schemas/PhysicalData" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Data not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "summary": "Delete physical data by ID",
        "tags": ["Physical Data"],
        "operationId": "deletePhysicalData",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Physical data ID",
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Physical data deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "example": "Success" },
                    "message": {
                      "type": "string",
                      "example": "Data deleted successfully"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Data not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "PhysicalData": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "userId": { "type": "string" },
          "weight": { "type": "number", "format": "float" },
          "height": { "type": "number", "format": "float" },
          "daily_activity": { "type": "string" },
          "target_calory": { "type": "number", "format": "float" },
          "target_protein": { "type": "number", "format": "float" },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" }
        }
      },
      "PhysicalDataInput": {
        "type": "object",
        "properties": {
          "weight": { "type": "number", "format": "float", "example": 70.5 },
          "height": { "type": "number", "format": "float", "example": 175.2 },
          "daily_activity": { "type": "string", "example": "Moderate" }
        },
        "required": ["weight", "height", "daily_activity"]
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}